+++ sources/generic/ft_delta_time.c
@@ -6,16 +6,16 @@
 /*   By: faru <faru@student.codam.nl>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/07/21 11:37:18 by faru          #+#    #+#                 */
-/*   Updated: 2023/07/21 11:38:29 by faru          ########   odam.nl         */
+/*   Updated: 2023/07/23 15:11:58 by fra           ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-long	ft_delta_time(t_timeval t1, t_timeval t2)
+double	ft_delta_time(t_timeval t1, t_timeval t2)
 {
 	t_timeval	delta;
-	long		timestamp;
+	double		timestamp;
 
 	while ((t2.tv_usec < t1.tv_usec) && (t2.tv_sec > t1.tv_sec))
 	{
@@ -25,5 +25,5 @@ long	ft_delta_time(t_timeval t1, t_timeval t2)
 	delta.tv_sec = t2.tv_sec - t1.tv_sec;
 	delta.tv_usec = t2.tv_usec - t1.tv_usec;
 	timestamp = delta.tv_sec * 1000000 + delta.tv_usec;
-	return (timestamp / 1000);
+	return (timestamp / 1000.);
 }
