+++ sources/numbers/ft_part_dec.c
@@ -6,7 +6,7 @@
 /*   By: fra <fra@student.codam.nl>                   +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/03/06 07:08:25 by fra           #+#    #+#                 */
-/*   Updated: 2023/07/21 23:10:14 by fra           ########   odam.nl         */
+/*   Updated: 2023/07/23 15:33:02 by fra           ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 double	ft_part_dec(double number)
 {
 	if (number > 0)
-		return (number - ft_part_int(number));
+		return ((double) number - ft_part_int(number));
 	else
-		return (number - ft_part_int(number) - 1);
+		return ((double) number - ft_part_int(number) - 1);
 }
